+++ app/src/main/java/de/baumann/weather/Browser.java
@@ -30,6 +30,8 @@ import android.view.ContextMenu;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+import android.webkit.DownloadListener;
+import android.webkit.URLUtil;
 import android.webkit.WebChromeClient;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
@@ -134,6 +136,7 @@ public class Browser extends AppCompatActivity  {
         mWebView.getSettings().setCacheMode(WebSettings.LOAD_DEFAULT); // load online by default
         mWebView.getSettings().setBuiltInZoomControls(true);
         mWebView.getSettings().setDisplayZoomControls(false);
+        registerForContextMenu(mWebView);
 
         if (sharedPref.getBoolean ("java", false)){
             mWebView.getSettings().setJavaScriptEnabled(true);
@@ -178,7 +181,34 @@ public class Browser extends AppCompatActivity  {
             }
         });
 
-        registerForContextMenu(mWebView);
+        mWebView.setDownloadListener(new DownloadListener() {
+
+            public void onDownloadStart(final String url, String userAgent,
+                                        final String contentDisposition, final String mimetype,
+                                        long contentLength) {
+
+                final String filename= URLUtil.guessFileName(url, contentDisposition, mimetype);
+                Snackbar snackbar = Snackbar
+                        .make(mWebView, getString(R.string.toast_download_1) + " " + filename, Snackbar.LENGTH_INDEFINITE)
+                        .setAction(getString(R.string.yes), new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                DownloadManager.Request request = new DownloadManager.Request(
+                                        Uri.parse(url));
+
+                                request.allowScanningByMediaScanner();
+                                request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED); //Notify client once download is completed!
+                                request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, filename);
+                                DownloadManager dm = (DownloadManager) getSystemService(DOWNLOAD_SERVICE);
+                                dm.enqueue(request);
+
+                                Snackbar.make(mWebView, getString(R.string.toast_download) + " " +
+                                        filename , Snackbar.LENGTH_LONG).show();
+                            }
+                        });
+                snackbar.show();
+            }
+        });
 
     }
 
@@ -243,7 +273,7 @@ public class Browser extends AppCompatActivity  {
                                     + source.getLastPathSegment());
                             request.setDestinationUri(Uri.fromFile(destinationFile));
                             ((DownloadManager) Browser.this.getSystemService(Context.DOWNLOAD_SERVICE)).enqueue(request);
-                            Snackbar.make(mWebView, R.string.context_saveImage_toast + " " +
+                            Snackbar.make(mWebView, getString(R.string.context_saveImage_toast) + " " +
                                     destinationFile.getAbsolutePath() , Snackbar.LENGTH_LONG).show();
                         }
                     }