+++ app/src/main/java/de/baumann/weather/helper/CustomListAdapter.java
@@ -32,6 +32,7 @@ public class CustomListAdapter extends ArrayAdapter<String> {
 
     public View getView(int position,View view,ViewGroup parent) {
         LayoutInflater inflater=context.getLayoutInflater();
+
         View rowView=inflater.inflate(R.layout.list_item, null,true);
 
         TextView txtTitle = (TextView) rowView.findViewById(R.id.item);